+++ scripts/dom.js
@@ -1295,6 +1295,19 @@ var readingActive = false, skipNextComic = false, skipPreviousComic = false;
 async function openComic(animation = true, path = true, mainPath = true, end = false, fromGoBack = false, fromNextAndPrev = false, fromDeepLoad = false)
 {
 	// Start reading comic
+	if(config.readingStartReadingInFullScreen && !fromNextAndPrev && !fromGoBack)
+	{
+		let win = electronRemote.getCurrentWindow();
+		let isFullScreen = win.isFullScreen();
+
+		if(!isFullScreen)
+		{
+			reading.hideContent(!isFullScreen);
+			win.setFullScreen(!isFullScreen);
+			win.setMenuBarVisibility(isFullScreen);
+		}
+	}
+
 	currentPathScrollTop[currentPath === false ? 0 : currentPath] = template.contentRight().children().scrollTop();
 	currentPath = path;